# docs/api.yaml
openapi: 3.0.0
info:
  title: Holiday Plans API
  description: API documentation for managing holiday plans
  version: 1.0.0
...
paths:
...
  /api/v1/holiday-plans/{holidayPlan}:
    get:
      summary: Get a holiday plan by ID
      operationId: show
      tags:
        - Holidays
      parameters:
        - name: holidayPlan
          in: path
          required: true
          description: ID of the holiday plan to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                id: 1
                title: test
                description: description
                date: 2024-03-16
                location: place
                participants:
                  - john
                  - doe
                url: http://example.com/api/v1/holiday-plans/1
        '404':
          description: Not found
    put:
      summary: Update a holiday plan
      operationId: update
      tags:
        - Holidays
      parameters:
        - name: holidayPlan
          in: path
          required: true
          description: ID of the holiday plan to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                date:
                  type: string
                  format: date
                location:
                  type: string
                participants:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal server
    delete:
      summary: Delete a holiday plan
      operationId: destroy
      tags:
        - Holidays
      parameters:
        - name: holidayPlan
          in: path
          required: true
          description: ID of the holiday plan to delete
          schema:
            type: integer
      responses:
        '204':
          description: No content
        '404':
          description: Not found
        '500':
          description: Internal server
...